+++ cmp/grid/impl/GridPersistenceModel.js
@@ -2,7 +2,7 @@
  * This file belongs to Hoist, an application development toolkit
  * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
  *
- * Copyright © 2020 Extremely Heavy Industries Inc.
+ * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 import {HoistModel, XH, managed, PersistenceProvider} from '@xh/hoist/core';
 import {observable, action, makeObservable} from '@xh/hoist/mobx';