+++ mobx/index.js
@@ -2,26 +2,27 @@
  * This file belongs to Hoist, an application development toolkit
  * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
  *
- * Copyright © 2020 Extremely Heavy Industries Inc.
+ * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 import {configure} from 'mobx';
 
 export {
-    extendObservable,
-    makeObservable,
     action,
     autorun,
+    comparer,
     computed,
+    extendObservable,
+    makeObservable,
     observable,
+    override,
     reaction,
     runInAction,
     toJS,
     trace,
     untracked,
-    when,
-    comparer
+    when
 } from 'mobx';
 export {observer} from 'mobx-react';
-export {settable, bindable} from './decorators';
+export {bindable, settable} from './decorators';
 
 configure({enforceActions: 'observed'});